requirements.txt
pandas
bs4
PyYAML
openpyxl